Add PSA manifest files for partition description.
This is a prototyping of PSA partition description in XML files
in the current uVisor version.
The current implementation recurcsively searches a given folder for
XML files starting with "box_" prefix which are treated as the manifest
files.
The corresponding code templates are expected to be located in "templates"
folder.
The script that does the job is generate_partition_code.py.